move agent id to env and move conversation component to components folder
